requirements.txt
fastapi==0.83.0
pydantic==1.10.2
pydantic[dotenv]
uvicorn==0.18.3
requests==2.28.1
SQLAlchemy==1.4.41
PyMySQL==1.0.2
cryptography==38.0.1
passlib==1.7.4
passlib[bcrypt]
python-jose==3.3.0 # JWT
python-jose[cryptography]